Add files via upload
I have tried to make this NumberGuessingGame a little bit different by adding GUI and APPLETS to the Program. 
The program uses two frames one is for rules and regulations and other is for the Game interface. 
The Game interface frame will have two buttons one to generate the number and other to guess the number.
It will also have a text field to tell whether the entered number is greater or lower to the computer generated number.
This game has 10 chances to guess the number correctly and the prize is 10000 $.
Each every chance you lose will decrease prize money by 1000 $.
I have included play again button also to play as many time you want.